Moved around the Signout function
