+++ src/App.js
@@ -11,12 +11,12 @@ import { Auth } from './components/Auth.js';
 import { ConversationList } from './components/ConversationList.js';
 import { Chat } from "./components/Chat.js"
 import { NewChat } from './components/newChat.js'
-import { Layout } from  "./Layout.js"
+import { Layout } from "./components/SignOut.js"
 
 const cookies = new Cookies(); //get, set, and remove cookies from browser
 
 
-function App() { 
+function App() {
   const [isAuth, setIsAuth] = useState(cookies.get("auth-token")) //if there is an auth-token then set to true (you can check by inspecting page manually)
 
   //once a user logs in, uuidv4 will generate a unique room ID
@@ -24,49 +24,49 @@ function App() {
 
   const roomInputRef = useRef(null)
 
-  const signUserOut = async () => {
-    try {
-      await signOut(auth);
-      cookies.remove("auth-token");
-      setIsAuth(false);
-      setRoom(null);
-    } catch (error) {
-      console.error('Error signing out:', error);
-    }
-  };
+  // const signUserOut = async () => {
+  //   try {
+  //     await signOut(auth);
+  //     cookies.remove("auth-token");
+  //     setIsAuth(false);
+  //     setRoom(null);
+  //   } catch (error) {
+  //     console.error('Error signing out:', error);
+  //   }
+  // };
 
   if (!isAuth) { //user is not authenticated
     window.location.href = "/auth";
   }
-  
+
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<Layout signUserOut={signUserOut} setIsAuth={setIsAuth} isAuth={isAuth}/>}>
-          <Route path="/auth" element={<Auth setIsAuth = {setIsAuth} />} />
-          <Route path='/new-chat' element={<NewChat />} />
-          <Route path="/chat/:room" element={<Chat isAuth={isAuth}/>} />
-          <Route path="/conversation-list" element={<ConversationList setRoom={setRoom} />} />
-        </Route>
+        {/* <Route path="/" element={<Layout signUserOuty={signUserOut} setIsAuthh={setIsAuth} isAuthh={isAuth} />}> */}
+        <Route path="/auth" element={<Auth setIsAuth={setIsAuth} />} />
+        <Route path='/new-chat' element={<NewChat />} />
+        <Route path="/chat/:room" element={<Chat isAuth={isAuth} />} />
+        <Route path="/conversation-list" element={<ConversationList setRoom={setRoom} />} />
+
       </Routes>
-    </BrowserRouter>
+    </BrowserRouter >
   );
 }
 
-  
 
-    //IMPLEMENTATION BELOW: POSSIBLE ROOM BEFORE ENTERING CHAT
-  // return  <div> { room ? ( //if user is authenticated, redircts to chat
-  //     <Chat room = {"NormChat"} /> //passes room variable through Chat component
 
-  //     ) : (
-  //       <div className = "room"> 
-  //         <label> Topic </label> 
-  //         <input  ref = {roomInputRef}/> {/**every chat has a reference assigned to it through "ref" */}
-  //         <button onClick = {() => setRoom(roomInputRef.current.value)}> Enter Chat </button>  {/**getting value of input and setting it to value of input */}
-  //       </div>
-  //     )}
-  //   </div>;
+//IMPLEMENTATION BELOW: POSSIBLE ROOM BEFORE ENTERING CHAT
+// return  <div> { room ? ( //if user is authenticated, redircts to chat
+//     <Chat room = {"NormChat"} /> //passes room variable through Chat component
+
+//     ) : (
+//       <div className = "room"> 
+//         <label> Topic </label> 
+//         <input  ref = {roomInputRef}/> {/**every chat has a reference assigned to it through "ref" */}
+//         <button onClick = {() => setRoom(roomInputRef.current.value)}> Enter Chat </button>  {/**getting value of input and setting it to value of input */}
+//       </div>
+//     )}
+//   </div>;
